SimpleAgent/requirements.txt
crewai
crewai-tools
#duckduckgo-search
python-dotenv
langchain-community
langchain_openai
streamlit
#youtube-transcript-api
#youtube-dl
pillow 
requests